requirements.txt
streamlit==1.44.1
ultralytics==8.3.140
opencv-python-headless==4.11.0.86
pillow==11.2.1
numpy==1.26.4
torch==2.6.0
torchvision==0.21.0

